Wrote the test case to validate calculateTimeBased_SpeedDistanceTravelled
The steps followed were based on

1. we know the distance to be travelled

2. to run in isolation. Distance is constant and so is

We would know the time it takes for a car to finish the distance we give it dist = speed x time

So assuming it takes 16 seconds.

For loop, we divide it by 4 for the time we pass it. aka it'll be in 4 steps for each time we do distance and speed calculation in calculateTimeBased_SpeedDistanceTravelled

To validate the above, the steps are:

instantiate a car variable and test if things are correctly assigned correctly

- doesn't finish as soon as you call it. Unless, distance is reached, it isn't finished.

- current speed is calculation correctly compared like we did before

- similar for distance